feat: add property area search using query parms
If a 'postcode' and 'radius' (in miles) are supplied as query
parameters, they are verified and 'postcode' geocode information is
fetched. This information is then used to calculate minimum and maximum
longitudes and latitudes (bounding box), which is used to filter
property objects and create the search results queryset.

Radius conversion to float and a custom exception were created to verify
the query parameter.