fix: add error handling to property update
Wrapped the code within the perform_update method in a try/except block
to account partial updates (PATCH) where postcode information is not
included.

Postcodes are checked when a record is updated to ensure the longitude
and latitude is updated (if required). This worked for PUT requests but
partial updates (PATCH) had not initially been accounted for.

Identified during automated testing.