fix: create permission class for note list view
A permissions class has been created which allows access to
authenticated users but only allows SAFE (read-only) request methods to
be used by Anonymous users.

Previously, anonymous users would receive a 403 error when attempting to
retrieve notes. This caused an unforeseen problem in the accompanying
frontend application  when trying to group API calls as the request
would fail when the 403 error was encountered (promise failure). The
simplest solution was a permissions change in the API to deny creation
of a note, but allow the retrieval of a blank array as originally
intended.