fix: add profile owner id to serializer fields
This fix reflects changes required to make the frontend application more
robust.

Problem: When making a request to follow a sellers profile, the sellers
"user id" is required. In the frontend application I had planned to use
the sellers "profile id" when making the post request to follow a user
as they should share the same id number in an ideal scenario.

Initial testing had caused the id numbers of these tables to fall out of
sync so the "profile id" could not be relied upon to match the "user
id".

Solution: This change returns the profile owners id as part of the
profile information which can reliably be used to create a follow
request. If this were to happen due to user moderation/manipulation by
the DB admin in a production environment, this will no longer
cause a difficult to diagnose 400 error.